agentic_rag_streamlit/backend/requirements.txt
Flask
flask-cors
faiss-cpu
sentence-transformers
pandas
python-pptx
python-docx
PyPDF2